Update Collector example
