+++ go/opentelemetry/collector/client/Dockerfile
