+++ go/opentelemetry/collector/client/client.go
@@ -2,23 +2,20 @@
 // example code to test lightstep/opentelemetry-exporter-go
 //
 // usage:
-//   go run server.go
+//   go run client.go
 
 package main
 
 import (
 	"context"
 	"fmt"
-	"log"
-	"math/rand"
-	"net/http"
-	"strings"
+	"io/ioutil"
+	// "net/http"
 	"time"
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/sdk/resource"
@@ -29,19 +26,11 @@ import (
 
 var (
 	tracer         trace.Tracer
-	serviceName    string = "test-go-server"
+	serviceName    string = "test-go-client-collector"
 	serviceVersion string = "0.1.0"
 	collectorAddr  string = "localhost:4318" // HTTP endpoint for collector
+	targetURL      string = "http://localhost:8081/ping"
 )
-
-const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-const (
-	letterIdxBits = 6                    // 6 bits to represent a letter index
-	letterIdxMask = 1<<letterIdxBits - 1 // All 1-bits, as many as letterIdxBits
-	letterIdxMax  = 63 / letterIdxBits   // # of letter indices fitting in 63 bits
-)
-
-var src = rand.NewSource(time.Now().UnixNano())
 
 func newTraceProvider(ctx context.Context) *sdktrace.TracerProvider {
 	exporter, err :=
@@ -85,54 +74,26 @@ func newTraceProvider(ctx context.Context) *sdktrace.TracerProvider {
 	)
 }
 
-func randString(n int) string {
-	sb := strings.Builder{}
-	sb.Grow(n)
-	for i, cache, remain := n-1, src.Int63(), letterIdxMax; i >= 0; {
-		if remain == 0 {
-			cache, remain = src.Int63(), letterIdxMax
-		}
-		if idx := int(cache & letterIdxMask); idx < len(letterBytes) {
-			sb.WriteByte(letterBytes[idx])
-			i--
-		}
-		cache >>= letterIdxBits
-		remain--
-	}
-
-	return sb.String()
-}
-
-// Wrap the handleRollDice so that telemetry data
-// can be automatically generated for it
-func wrapHandler() {
-	handler := http.HandlerFunc(handlePing)
-	wrappedHandler := otelhttp.NewHandler(handler, "pingHandler")
-	http.Handle("/ping", wrappedHandler)
-}
-
-func handlePing(w http.ResponseWriter, r *http.Request) {
-	operationName := "ping"
-	_, span := tracer.Start(r.Context(), operationName)
+func makeRequest(ctx context.Context) {
+	ctx, span := tracer.Start(ctx, "makeRequest")
 	defer span.End()
 
-	length := rand.Intn(1024)
-	log.Printf("%s %s %s", r.Method, r.URL.Path, r.Proto)
+	span.AddEvent("Making a request")
+	res, err := otelhttp.Get(ctx, targetURL)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer res.Body.Close()
+	body, _ := ioutil.ReadAll(res.Body)
+	fmt.Printf("Body : %s", body)
+	fmt.Printf("Request to %s, got %d bytes\n", targetURL, res.ContentLength)
 
-	pingResult := randString(length)
 	span.SetAttributes(
-		// attribute.String("result", pingResult),
-		attribute.String("library.language", "go"),
-		attribute.String("library.version", "v1.7.0"),
+		attribute.String("response", string(body)),
 	)
 
-	// setting span as successful
-	span.SetStatus(codes.Ok, "Success")
-
-	// setting span event
-	span.AddEvent(fmt.Sprint(r.Header))
-
-	fmt.Fprintf(w, pingResult)
+	span.AddEvent("Made a request", trace.WithAttributes(attribute.String("greeting", "Hello"), attribute.String("farewell", "Bye")))
 }
 
 func main() {
@@ -155,10 +116,9 @@ func main() {
 
 	tracer = tp.Tracer(serviceName, trace.WithInstrumentationVersion(serviceVersion))
 
-	wrapHandler()
-
-	err := http.ListenAndServe(":8081", nil)
-	if err != nil {
-		log.Fatal(err)
+	for {
+		makeRequest(ctx)
+		time.Sleep(1 * time.Second)
 	}
+
 }
