+++ go/opentelemetry/collector/server/Dockerfile
