+++ go/opentelemetry/collector/server/server.go
@@ -2,20 +2,23 @@
 // example code to test lightstep/opentelemetry-exporter-go
 //
 // usage:
-//   go run client.go
+//   go run server.go
 
 package main
 
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
-	// "net/http"
+	"log"
+	"math/rand"
+	"net/http"
+	"strings"
 	"time"
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/sdk/resource"
@@ -26,11 +29,19 @@ import (
 
 var (
 	tracer         trace.Tracer
-	serviceName    string = "test-go-client"
+	serviceName    string = "test-go-server-collector"
 	serviceVersion string = "0.1.0"
 	collectorAddr  string = "localhost:4318" // HTTP endpoint for collector
-	targetURL      string = "http://localhost:8081/ping"
 )
+
+const letterBytes = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
+const (
+	letterIdxBits = 6                    // 6 bits to represent a letter index
+	letterIdxMask = 1<<letterIdxBits - 1 // All 1-bits, as many as letterIdxBits
+	letterIdxMax  = 63 / letterIdxBits   // # of letter indices fitting in 63 bits
+)
+
+var src = rand.NewSource(time.Now().UnixNano())
 
 func newTraceProvider(ctx context.Context) *sdktrace.TracerProvider {
 	exporter, err :=
@@ -74,23 +85,54 @@ func newTraceProvider(ctx context.Context) *sdktrace.TracerProvider {
 	)
 }
 
-func makeRequest(ctx context.Context) {
-	ctx, span := tracer.Start(ctx, "makeRequest")
-	defer span.End()
-	res, err := otelhttp.Get(ctx, targetURL)
-	if err != nil {
-		fmt.Println(err)
-		return
+func randString(n int) string {
+	sb := strings.Builder{}
+	sb.Grow(n)
+	for i, cache, remain := n-1, src.Int63(), letterIdxMax; i >= 0; {
+		if remain == 0 {
+			cache, remain = src.Int63(), letterIdxMax
+		}
+		if idx := int(cache & letterIdxMask); idx < len(letterBytes) {
+			sb.WriteByte(letterBytes[idx])
+			i--
+		}
+		cache >>= letterIdxBits
+		remain--
 	}
-	defer res.Body.Close()
-	body, _ := ioutil.ReadAll(res.Body)
-	fmt.Printf("Body : %s", body)
-	fmt.Printf("Request to %s, got %d bytes\n", targetURL, res.ContentLength)
 
+	return sb.String()
+}
+
+// Wrap the handleRollDice so that telemetry data
+// can be automatically generated for it
+func wrapHandler() {
+	handler := http.HandlerFunc(handlePing)
+	wrappedHandler := otelhttp.NewHandler(handler, "pingHandler")
+	http.Handle("/ping", wrappedHandler)
+}
+
+func handlePing(w http.ResponseWriter, r *http.Request) {
+	operationName := "ping"
+	_, span := tracer.Start(r.Context(), operationName)
+	defer span.End()
+
+	length := rand.Intn(1024)
+	log.Printf("%s %s %s", r.Method, r.URL.Path, r.Proto)
+
+	pingResult := randString(length)
 	span.SetAttributes(
-		attribute.String("response", string(body)),
+		// attribute.String("result", pingResult),
+		attribute.String("library.language", "go"),
+		attribute.String("library.version", "v1.7.0"),
 	)
 
+	// setting span as successful
+	span.SetStatus(codes.Ok, "Success")
+
+	// setting span event
+	span.AddEvent(fmt.Sprint(r.Header))
+
+	fmt.Fprintf(w, pingResult)
 }
 
 func main() {
@@ -113,9 +155,10 @@ func main() {
 
 	tracer = tp.Tracer(serviceName, trace.WithInstrumentationVersion(serviceVersion))
 
-	for {
-		makeRequest(ctx)
-		time.Sleep(1 * time.Second)
-	}
+	wrapHandler()
 
+	err := http.ListenAndServe(":8081", nil)
+	if err != nil {
+		log.Fatal(err)
+	}
 }
